pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ffengine"
version = "0.1.0"
description = "Fast Flow Engine Community: Query-centric Airflow Native Data Integrator"
authors = [
    {name = "FFEngine Team", email = "info@ffengine.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.6",
    "psycopg[binary]>=3.1.0",
    "pyodbc>=5.0.0",
    "oracledb>=2.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "flake8>=6.0",
    "black>=23.0",
    "httpx>=0.27.0",
    "fastapi>=0.116.0,<0.118.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.pytest.ini_options]
markers = [
    "integration: container-backed integration tests",
]
